Prevent invalid anon function generation
Consider the user as having left the as soon as there's any non-whitespace text changes. It's a bit on the defensive side (sometimes it can be annoying that you need to re-trigger signature completion to get the code action again), but on the upside it should be impossible to get it to generate invalid code.