Make sure compilation requests are non-cancellable
To achieve this, server restarts are now delayed until all non-cancellable requests have been processed.

This issue could occur in the opposite case as the one in the previous commit: if a compilation has been started, and something leads to a server restart, the compilation would be cancelled halfway through (and the build process would just hang).