Format types in a way that plays more nicely with VSCode's sorting
It's important that the unqualified type name comes first, otherwise "Json" won't be the first match for "haxe.Json" (matches at the start seem to be weighted more strongly). With "Json - haxe" (unqualified name - container), we seem to get the desired sorting. This is inspired by the Java extension / language server.